Starting Place
Have written up a basic version of the interface as well as two
versions of the code. One utilizing JQuery, another handles the AJAX
calls with just Javascript. Still haven’t decided exactly where I’m
going with this, but for the moment, I have a starting place that is
somewhat scalable.