Simple interface now setup
A few bugs left to fix in layout spacing and stuff, but I have a search
view and a result view that are working.

Also, while I think my concept if a single field search with RegEx to
parse out dates vs. titles, What if a movie has a date as its title?