Added a basic implementation of pushState
Allows a user to use the back button to return to a prior search. Have
NOT QA’ed this fully and I’m certain there are still bugs in this. So
far, I’m fairly certain it will work to allow a user to go from a
specific choice (out of a search) back to the page of search results.
(Which you’d think would be the most likely use anyway).