Updated pass-to-controller method
While for the sake of just a cool implementation I liked using
setInterval, it ultimately isn’t necessary at this point as we’re
waiting on just one promise event that is already being handled
asynchronously and is capable of triggering the view controller all on
its own.

The issue was, I still didn’t want to include a direct call to
“outputData” in the omdbSearch object’s searchAPI method since to me,
that violates MVC ideas by placing responsibility for view output on
the model.

However, by adding a single argument to the method searchAPI() so it is
now searchAPI(callBack), the method remains abstract and flexible—I can
dictate which view controller is called when the XML request is
complete without having to modify the object code for different
implementations. For now, I think it’s a decent way to go.

TODO:
Implement a global run loop of sorts (similar to NSRunLoop in Obj-C)
that would begin execution upon page load. I’d feed it a object and a
state (in this case MovieSearcher.bSearchComplete == true), and when
that state is achieved, tell it to call a specific function or object
method to begin display of the data. Ultimately, that’s the abstraction
I want here, just don’t have the time (or perhaps the need) to
implement right now.