RunLoop Experiment + SearchVC Stub
Woke up with a thought about my Run Loop. Decided to test.

I’m now using setInterval (in a global context, executed on page load)
to create my own Run Loop. Implemented a very simple version of it here
just so I could see how it affected responsiveness in the browser. So
far, I’ve tested intervals down to 100ms without any noticeable
degradation (at least no noticeable degradation in this very simple web
app). All this Run Loop does (for the moment) is continually check on
the state of MovieSearcher.bSearchComplete and send a message to the
console indicating it has achieved its goal. Since this does what its
intended to do so far, I’m going to see if I can ultimately use it to
avoid having to send a specific callback function to the
omdbSearch.searchAPI() method as this would allow for a near complete
distinction between the model and view controller. The Run Loop listens
for a designated state change, and when it happens, the execution flow
continues.

TODO:
1. Remove callBack from searchAPI method and update code to call view
controller using this event listener instead.

2. Abstract this idea into a class that is essentially a private
(non-DOM) event listener. Need to be able to register events to listen
for in an organized way and then take specific actions.

——

Also added a stub for a SearchViewController that I’ll implement at
some point.